requirements.txt
# ── NLP ─────────────────────────────────────────────────────────────────────
# NOTE: spaCy is NOT listed — it is incompatible with Python 3.14+
# RAD-ML v7 uses pure-Python NLP (stdlib only) in brain/prompt_parser.py
nltk>=3.8.0

# ── Data collection ──────────────────────────────────────────────────────────
kaggle>=1.6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# ── Data processing ───────────────────────────────────────────────────────────
numpy>=1.26.0
scikit-learn>=1.4.0
pandas>=2.2.0

# ── AWS ───────────────────────────────────────────────────────────────────────
boto3>=1.34.0
sagemaker>=2.200.0

# ── LLM ───────────────────────────────────────────────────────────────────────
google-generativeai>=0.7.0

# ── Web / API server ──────────────────────────────────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0

# ── Config ────────────────────────────────────────────────────────────────────
pyyaml>=6.0.0
python-dotenv>=1.0.0
rich>=13.7.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0

# ── Auth & database ───────────────────────────────────────────────────────────
bcrypt>=4.1.0
google-auth>=2.25.0

# ── OpenML ────────────────────────────────────────────────────────────────────
openml>=0.15.0

# ── HuggingFace ───────────────────────────────────────────────────────────────
huggingface_hub>=0.20.0
datasets>=2.14.0

# ── Explainability ────────────────────────────────────────────────────────────
diagrams>=0.23.0

# ── VM-ERA (Visual Malware Fingerprinting & Ensemble Risk Assessment) ─────────
# APK processing
androguard>=3.4.0a1
Pillow>=9.0.0

# Deep learning models (GPU-enabled for Windows)
tensorflow==2.10.1
protobuf<3.20

# Ensemble models
joblib>=1.2.0

# UI enhancements (optional)
plotly>=5.14.0
